refactor: move Seller page, LUC page from components into pages
