+++ src/pages/LightUpChinatown/images/goal_1.png
