+++ src/pages/LightUpChinatown/images/goal_2.png
