+++ src/pages/LightUpChinatown/images/lantern-hero-top.png
