+++ src/pages/LightUpChinatown/images/partners-cccny.png
