+++ src/pages/LightUpChinatown/images/partners-nyccc.png
